backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
python-dotenv==1.1.1
requests==2.32.4
readability-lxml==0.8.4.1
trafilatura==2.0.0
openai==1.99.9
itsdangerous==2.2.0
